Bug 933654 - 1/3: refactor amodem_send.
The original amodem_send returns the reply string of a AT command,
and then ModemDriver determines whether the modem is now waiting for
SMS by comparing that string to '> '.  However, as we're to enable
unsolicited responses after the reply, such method is no longer
effective and could be error prone.

This patch moves all reply handling into android_modem.c and returns
and modem->waiting_sms in the end of amodem_send().

Signed-off-by: Tony Espy <espy@canonical.com>